chore: move order table components to folder components
